pyproject.toml
[project]
name = "sparkopy"
version = "0.0.1"
authors = [
  { name="Ryan Duryea", email="aguynamedryan+oss@gmail.com" },
]
description = "Downloads tables from Spark and outputs them as Parquet files"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

dependencies = [
  "click==8.1.7",
  "databricks-connect==14.3.1",
  "grpcio",
  "grpcio-status",
  "pandas==2.2.2",
  "protobuf",
  "pyarrow==16.0.0"
]

[project.urls]
Homepage = "https://github.com/pypa/sampleproject"
Issues = "https://github.com/pypa/sampleproject/issues"

[project.scripts]
sparkopy = "sparkopy:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"